# A faster s2b function