--- a/bytesconv.go
+++ b/bytesconv.go
@@ -386,7 +386,7 @@ func b2s(b []byte) string {
 // in the future go versions.
 func s2b(s string) (b []byte) {
 	bh := (*reflect.SliceHeader)(unsafe.Pointer(&b))
-	sh := (*reflect.StringHeader)(unsafe.Pointer(&s))
+	sh := *(*reflect.StringHeader)(unsafe.Pointer(&s))
 	bh.Data = sh.Data
 	bh.Len = sh.Len
 	bh.Cap = sh.Len