@ZYunH 2019-08-19
I also think so before, but from the point of view of Go compiler ASM, keep sh as *(*reflect.StringHeader)(unsafe.Pointer(&s)) costing lower than keep it as a pointer. There may some compiler optimizations in this case. Let's see it in ASM.
(Go code)

> func s2bV1(s string) (b []byte) {
	bh := (*SliceHeader)(unsafe.Pointer(&b))
	sh := *(*StringHeader)(unsafe.Pointer(&s))
	bh.Data = sh.Data
	bh.Len = sh.Len
	bh.Cap = sh.Len
	return b
}

> func s2bV1(s string) (b []byte) {
	bh := (*SliceHeader)(unsafe.Pointer(&b))
	sh := *(*StringHeader)(unsafe.Pointer(&s))
	bh.Data = sh.Data
	bh.Len = sh.Len
	bh.Cap = sh.Len
	return b
}

(In ASM, without GC code)

> "".s2bV1 STEXT nosplit size=43 args=0x28 locals=0x0
	0x0000 00000 (t_main2.go:18)	TEXT	"".s2bV1(SB), NOSPLIT|ABIInternal, $0-40
	0x0011 00017 (t_main2.go:20)	MOVQ	"".s+16(SP), AX
	0x0016 00022 (t_main2.go:20)	MOVQ	"".s+8(SP), CX
	0x001b 00027 (t_main2.go:21)	MOVQ	CX, "".b+24(SP)
	0x0020 00032 (t_main2.go:22)	MOVQ	AX, "".b+32(SP)
	0x0025 00037 (t_main2.go:23)	MOVQ	AX, "".b+40(SP)
	0x002a 00042 (t_main2.go:24)	RET

> "".s2bV2 STEXT nosplit size=48 args=0x28 locals=0x0
	0x0000 00000 (t_main2.go:27)	TEXT	"".s2bV2(SB), NOSPLIT|ABIInternal, $0-40
	0x0011 00017 (t_main2.go:30)	MOVQ	"".s+8(SP), AX
	0x0016 00022 (t_main2.go:30)	MOVQ	AX, "".b+24(SP)
	0x001b 00027 (t_main2.go:31)	MOVQ	"".s+16(SP), AX
	0x0020 00032 (t_main2.go:31)	MOVQ	AX, "".b+32(SP)
	0x0025 00037 (t_main2.go:32)	MOVQ	"".s+16(SP), AX
	0x002a 00042 (t_main2.go:32)	MOVQ	AX, "".b+40(SP)
	0x002f 00047 (t_main2.go:33)	RET

You can see the version one can use more registers and version two can use only one, so V1 has fewer instructions in ASM, actually V1 use 5 instructions, V2 use 6 instructions. (Size of V1 is 43, and size of V2 is 48, and this is the only difference.)